deleted filesand renamed files
